Adding MBassador, refactoring game to outside of MVC
