+++ build.gradle
@@ -58,6 +58,7 @@ project(":core") {
         compile "com.badlogicgames.gdx:gdx-ai:1.8.0"
         compile "com.esotericsoftware:kryonet:2.22.0-RC1"
         compile "com.beust:jcommander:1.72"
+        compile "net.engio:mbassador:1.3.1"
     }
 }
