+++ core/src/com/keenant/secutor/engine/controller/Controller.java
@@ -19,5 +19,5 @@ public interface Controller<M extends Model, V extends View<M>> {
 
   void setView(V view);
 
-  void update(float deltaTime);
+  void act(float deltaTime);
 }
