+++ core/src/com/keenant/secutor/event/Event.java
@@ -1,0 +1,5 @@
+package com.keenant.secutor.event;
+
+public interface Event {
+
+}
